Pull out combined fits into fitting.py.  Make master fitting script
